move_rot

+++ test.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 #include <math.h>
 
-// void	DrawLine(t_data *data, float x_start, float y_start)
+// void	DrawLine(t_data *data, double x_start, double y_start)
 // {
-// 	float x_end = x_start + (cos(ANGLE * M_PI / 180) * LINE_LENGHT);
-// 	float y_end = y_start + (sin(ANGLE * M_PI / 180) * LINE_LENGHT);
-// 	float m;
-// 	float b;
+// 	double x_end = x_start + (cos(ANGLE * M_PI / 180) * LINE_LENGHT);
+// 	double y_end = y_start + (sin(ANGLE * M_PI / 180) * LINE_LENGHT);
+// 	double m;
+// 	double b;
 
 // 	// fprintf(stderr, "x_end == %f, y_end %f\n", x_end, y_end);
 // 	m = ((y_end - y_start) / (x_start - x_end));
 // 	b = y_start - (m * x_start);
 // 	fprintf(stderr, "m == %f, b == %f\n", m, b);
-// 	for (float y = y_start; y <= y_end; y += 0.1)
+// 	for (double y = y_start; y <= y_end; y += 0.1)
 // 	{
-// 		for(float x = x_start; x <= x_end; x += 0.1)
+// 		for(double x = x_start; x <= x_end; x += 0.1)
 // 		{
-// 			if (roundf((y - (m * x) - b)) == 0)
+// 			if (round((y - (m * x) - b)) == 0)
 // 			{
 // 				write(2, "HERE\n", 5);
 // 				fprintf(stderr, "x == %f, y == %f\n", x, y);
@@ -35,10 +35,10 @@
 
 int main()
 {
-	float a = 2.554;
-	float b = 2.123456;
+	double a = 2.554;
+	double b = 2.123456;
 
 
-	float r = remainderf(a, b);
+	double r = remainderf(a, b);
 	printf("remainder == %f\n", r);
 }
